% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{add.terms}
\alias{add.terms}
\title{Add terms to an lme4 formula.}
\usage{
add.terms(formula, add)
}
\arguments{
\item{formula}{The lme4 formula to add terms to.}

\item{add}{A vector of terms to add. To add terms nested in random-effect groups, use `(term|group)' syntax if you want to add an independent random effect (e.g. `(olderterm|group) + (term|group)'), or use `term|group' syntax if you want to add a dependent random effect to a pre-existing term group (if no such group exists, it will be created at the end of the formula).}
}
\value{
The updated formula.
}
\description{
Add terms to an lme4 formula.
}
\seealso{
buildmer
}
\keyword{add}
\keyword{terms}

