% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{build.formula}
\alias{build.formula}
\title{Convert buildmer term list into a proper model formula}
\usage{
build.formula(dep, terms)
}
\arguments{
\item{dep}{The dependent variable.}

\item{terms}{The term list.}
}
\value{
A formula.
}
\description{
Convert buildmer term list into a proper model formula
}
