% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular.R
\docType{class}
\name{buildmer-class}
\alias{buildmer-class}
\alias{mkBuildmer}
\title{Make a buildmer object}
\arguments{
\item{model}{The final model containing only the terms that survived elimination.}

\item{p}{Parameters used during the fitting process.}

\item{anova:}{The model's ANOVA, if the model was built with `anova=TRUE'.}

\item{summary:}{The model's summary, if the model was built with `summary=TRUE'.}
}
\description{
Make a buildmer object
}
\seealso{
buildmer
}
\keyword{buildmer}
