% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\docType{class}
\name{buildmer.terms-class}
\alias{buildmer.terms-class}
\alias{mkBMTerms}
\title{Make a buildmer terms object.}
\arguments{
\item{y}{The dependent variable.}

\item{terms}{A list consisting of two elements named 'fixed' and 'random', which contain fixed and random factors (respectively) as character strings. If 'random' is an empty list, the model will be fit using (g)lm, otherwise (g)lmer from the lme4 package will be used (or, if it is available, from the lmerTest package instead).}

\item{data}{The data to fit the model to.}

\item{data.name}{Internal parameter used by buildmer to replace the name of your dataset by the argument passed to buildmer (as opposed to just 'data').}

\item{family}{The error distribution to use. Only relevant for generalized models; ignored for linear models.}

\item{diag}{Whether to assume a diagonal covariance structure.}

\item{verbose}{The verbosity level passed to (g)lmer fits.}

\item{maxfun}{The maximum number of iterations to allow for (g)lmer fits.}
}
\description{
Make a buildmer terms object.
}
\keyword{terms}

