% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildmultinom}
\alias{buildmultinom}
\title{Use \code{buildmer} to perform stepwise elimination for \code{multinom} models from package \code{nnet}}
\usage{
buildmultinom(
  formula,
  data = NULL,
  cl = NULL,
  direction = c("order", "backward"),
  crit = "LRT",
  include = NULL,
  calc.summary = TRUE,
  ...
)
}
\arguments{
\item{formula}{See the general documentation under \code{\link{buildmer-package}}}

\item{data}{See the general documentation under \code{\link{buildmer-package}}}

\item{cl}{See the general documentation under \code{\link{buildmer-package}}}

\item{direction}{See the general documentation under \code{\link{buildmer-package}}}

\item{crit}{See the general documentation under \code{\link{buildmer-package}}}

\item{include}{See the general documentation under \code{\link{buildmer-package}}}

\item{calc.summary}{See the general documentation under \code{\link{buildmer-package}}}

\item{...}{Additional options to be passed to \code{multinom}}
}
\description{
Use \code{buildmer} to perform stepwise elimination for \code{multinom} models from package \code{nnet}
}
\examples{
if (requireNamespace('nnet') && require('MASS')) {
	options(contrasts = c("contr.treatment", "contr.poly"))
	example(birthwt)
	bwt.mu <- buildmultinom(low ~ age*lwt*race*smoke,bwt)
}
}
\seealso{
\code{\link{buildmer-package}}
}
