% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexFuns.R
\name{calcor}
\alias{calcor}
\title{Calculate Odds Ratio; basically prints either 'exp(x):1' or '1:exp(x)'}
\usage{
calcor(x)
}
\description{
Calculate Odds Ratio; basically prints either 'exp(x):1' or '1:exp(x)'
}
