% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{calcWald}
\alias{calcWald}
\title{Calculate p-values based on Wald z-scores}
\usage{
calcWald(table, i, sqrt = F)
}
\arguments{
\item{table}{A table from a summary or anova output.}

\item{i}{The number of the column in that table containing the t-values}

\item{sqrt}{Whether we're testing F values or t values (default)}
}
\value{
The table augmented with a column 'p (Wald)' containing Wald p values
}
\description{
Calculate p-values based on Wald z-scores
}

