% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{contr.deviation}
\alias{contr.deviation}
\title{Sets up deviation contrasts}
\usage{
contr.deviation(n, contrasts = TRUE, sparse = FALSE)
}
\arguments{
\item{n}{See \code{\link{contr.sum}}.}

\item{contrasts}{See \code{\link{contr.sum}}.}

\item{sparse}{See \code{\link{contr.sum}}.}
}
\value{
A matrix of deviation-coded contrasts.
}
\description{
Deviation-coding is a contrast-coding scheme that compares each level to the average of the other levels. It is often confused with sum-coding, which compares each level to the average of \emph{all} levels. In the binary case, deviation coding works out to \( (+0.5,-0.5) \); in the ternary case and beyond, the coding is a little bit more involved.
}
\details{
\code{contr.deviation} starts by calling \code{\link{contr.sum}} and hence accepts the same arguments as it does.
}
\examples{
iris$Species <- contr.deviation(iris$Species)
}
\seealso{
\code{\link{contr.sum}}
}
