% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular.R
\name{devfun}
\alias{devfun}
\title{Extract a model's deviance}
\usage{
devfun(model)
}
\arguments{
\item{model}{The fitted model object.}
}
\value{
The deviance or REML criterion.
}
\description{
Extract a model's deviance
}
