% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{deviation}
\alias{deviation}
\title{Sets up deviation contrasts}
\usage{
deviation(n, contrasts = TRUE, sparse = FALSE)
}
\arguments{
\item{n}{See \code{\link{contr.sum}}.}

\item{contrasts}{See \code{\link{contr.sum}}.}

\item{sparse}{See \code{\link{contr.sum}}.}
}
\value{
A matrix of deviation-coded contrasts.
}
\description{
This function is an alias to \code{\link{contr.deviation}} intended to allow the \code{C} command to work for deviation-coding the same way as it does for other, base-R, contrasts.
}
\examples{
iris$Species <- C(iris$Species,deviation)
}
\seealso{
\code{\link{contr.deviation}}, \code{\link{contr.sum}}
}
