% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{diagonalize}
\alias{diagonalize}
\title{Diagonalize the random-effect covariance structure, possibly assisting convergence}
\usage{
diagonalize(formula)
}
\arguments{
\item{formula}{A model formula.}
}
\value{
The formula with all random-effect correlations forced to zero, per Pinheiro & Bates (2000).
}
\description{
Diagonalize the random-effect covariance structure, possibly assisting convergence
}

