% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{elim}
\alias{elim}
\title{Get the elimination table from a buildmer object}
\usage{
elim(object)
}
\arguments{
\item{object}{The buildmer object.}
}
\description{
Get the elimination table from a buildmer object
}
\seealso{
buildmer
}
\keyword{elimination}

