% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{fit}
\alias{fit}
\title{Fit a model using either lme4 or (g)lm, with formula terms passed via a buildmer.terms object.}
\usage{
fit(bmt, REML = TRUE)
}
\arguments{
\item{bmt}{The buildmer.terms object containing the predictors, data, and fitting parameters. See buildmer.terms.}

\item{REML}{Whether to fit the model using REML (default) or ML. Only relevant for linear mixed effects models; ignored for other models.}
}
\description{
Fit a model using either lme4 or (g)lm, with formula terms passed via a buildmer.terms object.
}
\seealso{
buildmer, mkBMTerms
}
\keyword{fit}

