% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{get.last.random.slope}
\alias{get.last.random.slope}
\title{Get the last random slope (or, if not available, the last random intercept) from a model formula}
\usage{
get.last.random.slope(formula)
}
\arguments{
\item{formula}{A model formula.}
}
\value{
The last random slope (or, if not available, the last random intercept).
}
\description{
Get the last random slope (or, if not available, the last random intercept) from a model formula
}

