% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{get.random.terms}
\alias{get.random.terms}
\title{Get the level-2 terms contained within a level-1 lme4 random term}
\usage{
get.random.terms(term)
}
\arguments{
\item{term}{The term.}
}
\value{
A list of random terms.
}
\description{
Get the level-2 terms contained within a level-1 lme4 random term
}
