% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{has.smooth.terms}
\alias{has.smooth.terms}
\title{Test whether a formula contains gamm4 smooth terms}
\usage{
has.smooth.terms(formula)
}
\arguments{
\item{formula}{The formula.}
}
\value{
A logical indicating whether the formula has any gamm4 terms.
}
\description{
Test whether a formula contains gamm4 smooth terms
}

