% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{hasREML}
\alias{hasREML}
\title{Test whether a model was fit with REML}
\usage{
hasREML(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
TRUE or FALSE if the model was a linear mixed-effects model that was fit with REML or not, respectively; NA otherwise.
}
\description{
Test whether a model was fit with REML
}
