% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{innerapply}
\alias{innerapply}
\title{Convenience function to immediately descend into a level-2 term}
\usage{
innerapply(random.terms, FUN)
}
\arguments{
\item{random.terms}{Vector of level-1 terms to descend into.}

\item{FUN}{The function to apply to the level-2 terms.}
}
\value{
The modified random.terms
}
\description{
Convenience function to immediately descend into a level-2 term
}
