% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{is.random.term}
\alias{is.random.term}
\title{Test whether a formula term contains lme4 random terms}
\usage{
is.random.term(term)
}
\arguments{
\item{term}{The term.}
}
\value{
A logical indicating whether the term was a random-effects term.
}
\description{
Test whether a formula term contains lme4 random terms
}

