% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{is.smooth.term}
\alias{is.smooth.term}
\title{Test whether a formula term is an mgcv smooth term}
\usage{
is.smooth.term(term)
}
\arguments{
\item{term}{The term.}
}
\value{
A logical indicating whether the term was a random-effects term.
}
\description{
Test whether a formula term is an mgcv smooth term
}
