% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{mcmc2tex}
\alias{mcmc2tex}
\title{Convert an MCMCglmm model to LaTeX code (biased towards stress analysis)}
\usage{
mcmc2tex(model, aliases = list())
}
\arguments{
\item{model}{The fitted model}

\item{aliases}{A list of aliases translating summary terms to LaTeX code.}
}
\description{
Convert an MCMCglmm model to LaTeX code (biased towards stress analysis)
}
