% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{re2uncorr}
\alias{re2uncorr}
\title{Converts lme4 random-effect terms to uncorrelated lme4 random-effect terms, including factor random slopes}
\usage{
re2uncorr(formula, data, drop = TRUE)
}
\arguments{
\item{formula}{The lme4 formula.}

\item{data}{The data.}

\item{drop}{Logical indicating whether constant, non-intercept columns should be dropped. Default \code{TRUE}. A warning is issued if a column needed to be dropped. Note that repeated intercept columns are silently merged without a warning.}
}
\value{
A list containing a new formula (has a dependent variable but loses track of what terms belong together in a single block), a buildmer term list (lacks the dependent variable but represents what terms belong together), and a new data set.
}
\description{
lme4 and similar packages (e.g. glmmTMB) have an issue where correlations between \emph{factor} random slopes are not dropped, even if using the double-bar syntax or \code{\link{diag}} to diagonalize the random-effects formula. \code{re2uncorr} works around this issue by explicitly converting factor random slopes to numerics, which do not suffer from this problem.
}
\details{
To ensure that the various buildmer functions (e.g. buildmer, buildglmmTMB) correctly apply marginality constraints between fixed effects and random-effect terms, the fixed effects are renamed using the same name-mangling scheme that is automatically applied to the random effects. To ensure that the mangled names will be grouped together, instead of passing the returned \code{formula}, use the returned \code{termlist} with a \code{dep} argument. See the description of the \code{dep} argument in \code{\link{buildmerControl}}.
}
\examples{
library(buildmer)
re <- re2uncorr(f1 ~ vowel*timepoint*following +
	(vowel*timepoint*following|participant) + (timepoint|word),vowels)
\dontshow{if (FALSE)}
model <- buildmer(re$formula,re$data)
}
