% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{remove.terms}
\alias{remove.terms}
\title{Remove terms from an lme4 formula}
\usage{
remove.terms(formula, remove, formulize = T)
}
\arguments{
\item{formula}{The lme4 formula.}

\item{remove}{A vector of terms to remove. To remove terms nested inside random-effect groups, use `term|group' syntax. Note that marginality is respected, i.e. no effects will be removed if they participate in a higher-order interaction, and no fixed effects will be removed if a random slope is included over that fixed effect.}

\item{formulize}{Whether to return a formula (default) or a simple list of terms.}
}
\description{
Remove terms from an lme4 formula
}
\seealso{
buildmer
}
\keyword{remove}
\keyword{terms}
