% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{reorder.terms}
\alias{reorder.terms}
\title{Reorder the terms in a regression model by their contribution to the deviance (i.e. sort the terms by their significances)}
\usage{
\method{reorder}{terms}(formula, data, family = gaussian, nAGQ = 1,
  quiet = FALSE, verbose = 0, maxfun = 2e+05)
}
\arguments{
\item{formula}{The model formula; if you include random effects, use lme4 syntax for them.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or 'gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through.}

\item{nAGQ}{The number of Adaptive Gauss-Hermitian Quadrature points to use for glmer fits; default 1 (= Laplace approximation)}

\item{quiet}{Whether to suppress progress messages.}

\item{verbose}{The verbosity level passed to (g)lmer fits.}

\item{maxfun}{The maximum number of iterations to allow for (g)lmer fits.}
}
\value{
The formula, with the terms reordered.
}
\description{
Reorder the terms in a regression model by their contribution to the deviance (i.e. sort the terms by their significances)
}
\examples{
buildmer(Reaction~Days,list(Subject=~Days),lme4::sleepstudy)
}
\keyword{order}

