% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{stepwise}
\alias{stepwise}
\title{A simple interface to buildmer intended to mimic SPSS stepwise methods for term reordering and backward stepwise elimination}
\usage{
stepwise(formula, data, family = gaussian, ...)
}
\arguments{
\item{formula}{The model formula for the maximal model you would like to fit, if possible. Supports lme4 random effects and gamm4 smooth terms.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or 'gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through.}

\item{...}{Additional parameters that override buildmer defaults, see 'buildmer'.}
}
\value{
A buildmer object containing the following slots:
\itemize{
\item table: a dataframe containing the results of the elimination process
\item model: the final model containing only the terms that survived elimination
\item messages: any warning messages
\item summary: the model's summary, if 'summary=TRUE' was passed
\item anova: the model's anova, if 'anova=TRUE' was passed
}
}
\description{
A simple interface to buildmer intended to mimic SPSS stepwise methods for term reordering and backward stepwise elimination
}
\examples{
buildmer(Reaction~Days+(Days|Subject),sleepstudy)
}
\keyword{SPSS,}
\keyword{buildmer,}
\keyword{elimination,}
\keyword{order}
\keyword{stepwise}
\keyword{term}

