% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tabulate.formula}
\alias{tabulate.formula}
\title{Parse a formula into a buildmer terms list}
\usage{
tabulate.formula(formula)
}
\arguments{
\item{formula}{A formula.}
}
\value{
A buildmer terms list, which is just a normal data frame.
}
\description{
Parse a formula into a buildmer terms list
}
\examples{
form <- diag(f1 ~ vowel*timepoint*following + ((vowel1+vowel2+vowel3+vowel4)*timepoint*
                  following|participant) + (timepoint|word))
tabulate.formula(form)
}
