% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{unravel}
\alias{unravel}
\title{Unravel interaction terms (default) or additive term lists (specify sym='+')}
\usage{
unravel(x, sym = ":")
}
\arguments{
\item{x}{A terms object as obtained from terms(formula).}

\item{sym}{The symbol to split this terms object on. By default, this is ':', indicating that interactions should be split into their constituent terms. Set it to '+' to split a formula into its individual terms.}
}
\value{
A character vector of all the terms encountered.
}
\description{
Unravel interaction terms (default) or additive term lists (specify sym='+')
}
