% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\docType{data}
\name{vowels}
\alias{vowels}
\title{Vowel data from a pilot study.}
\format{A standard data frame.}
\usage{
data(vowels)
}
\description{
Vowel data from a pilot study.
}
\examples{
buildmer(f1 ~ vowel*timepoint*following + stress + information + (vowel*timepoint*following|participant),data=vowels)
}
\keyword{datasets}

