% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildbam}
\alias{buildbam}
\title{Use buildmer to fit big generalized additive models using bam() from mgcv}
\usage{
buildbam(formula, data, family = gaussian, reorder.terms = TRUE,
  cl = NULL, reduce.fixed = TRUE, reduce.random = TRUE,
  direction = "backward", crit = "LRT", calc.anova = TRUE,
  calc.summary = TRUE, ddf = "Wald", quiet = FALSE, ...)
}
\arguments{
\item{formula}{The model formula for the maximal model you would like to fit, if possible. Supports lme4 random effects and gamm4 smooth terms.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or `gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through.}

\item{reorder.terms}{Whether to reorder the terms by their contribution to the log-likelihood before testing them.}

\item{cl}{An optional cluster object as returned by parallel::makeCluster() to use for parallelizing the evaluation of terms during the reordering step.}

\item{reduce.fixed}{Whether to reduce the fixed-effect structure.}

\item{reduce.random}{Whether to reduce the random-effect structure.}

\item{direction}{The direction for stepwise elimination; either `forward' or `backward' (default). Both or neither are also understood.}

\item{crit}{The criterion used to test terms for elimination. Possible options are `LRT' (default), `AIC', and `BIC'.}

\item{calc.anova}{Whether to also calculate the ANOVA table for the final model after term elimination.}

\item{calc.summary}{Whether to also calculate the summary table for the final model after term elimination.}

\item{quiet}{Whether to suppress progress messages.}

\item{...}{Additional options to be passed to bam().}
}
\value{
A buildmer object containing the following slots:
\itemize{
\item model: the final model containing only the terms that survived elimination
\item p: the parameter list used in the various buildmer modules. Things of interest this list includes are, among others:
\itemize{
\item results: a dataframe containing the results of the elimination process
\item messages: any warning messages
}. This information is also printed as part of the show() method.
\item summary: the model's summary, if `calc.summary=TRUE' was passed
\item anova: the model's anova, if `calc.anova=TRUE' was passed
}
}
\description{
Use buildmer to fit big generalized additive models using bam() from mgcv
}
\seealso{
buildmer
}
