% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildgamm4}
\alias{buildgamm4}
\title{Use buildmer to fit generalized additive models using gamm4}
\usage{
buildgamm4(...)
}
\arguments{
\item{...}{Additional options to be passed to gam().}

\item{formula}{The model formula for the maximal model you would like to fit, if possible. Supports lme4 random effects and gamm4 smooth terms.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or `gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through.}

\item{cl}{An optional cluster object as returned by parallel::makeCluster() to use for parallelizing the evaluation of terms.}

\item{reduce.fixed}{Whether to reduce the fixed-effect structure.}

\item{reduce.random}{Whether to reduce the random-effect structure.}

\item{direction}{The direction for stepwise elimination; possible options are `order' (order terms by their contribution to the model), `backward' (backward elimination), `forward' (forward elimination, implies `order'). The default is the combination `c('order','backward')', to first make sure that the model converges and to then perform backward elimination; other such combinations are perfectly allowed.}

\item{crit}{The criterion used to test terms for elimination. Possible options are `LRT' (default), `AIC', and `BIC'.}

\item{calc.anova}{Whether to also calculate the ANOVA table for the final model after term elimination.}

\item{calc.summary}{Whether to also calculate the summary table for the final model after term elimination.}

\item{ddf}{The method used for calculating p-values if all smooth terms were eliminated and summary=TRUE. Options are `Wald' (default), `Satterthwaite' (if lmerTest is available), `Kenward-Roger' (if lmerTest and pbkrtest are available), and `lme4' (no p-values).}

\item{quiet}{Whether to suppress progress messages.}
}
\value{
A buildmer object containing the following slots:
\itemize{
\item model: the final model containing only the terms that survived elimination
\item p: the parameter list used in the various buildmer modules. Things of interest this list includes are, among others:
\itemize{
\item results: a dataframe containing the results of the elimination process
\item messages: any warning messages
} This information is also printed as part of the show() method.
\item summary: the model's summary, if `calc.summary=TRUE' was passed
\item anova: the model's anova, if `calc.anova=TRUE' was passed
}
}
\description{
Use buildmer to fit generalized additive models using gamm4
}
\seealso{
buildmer
}
