% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildgls}
\alias{buildgls}
\title{Use \code{buildmer} to fit generalized-least-squares models using \code{gls} from \code{nlme}}
\usage{
buildgls(formula, data = NULL, cl = NULL, direction = c("order",
  "backward"), crit = "LRT", include = NULL, calc.anova = TRUE,
  calc.summary = TRUE, ...)
}
\arguments{
\item{formula}{See the general documentation under \code{\link{buildmer-package}}}

\item{data}{See the general documentation under \code{\link{buildmer-package}}}

\item{cl}{See the general documentation under \code{\link{buildmer-package}}}

\item{direction}{See the general documentation under \code{\link{buildmer-package}}}

\item{crit}{See the general documentation under \code{\link{buildmer-package}}}

\item{include}{See the general documentation under \code{\link{buildmer-package}}}

\item{calc.anova}{See the general documentation under \code{\link{buildmer-package}}}

\item{calc.summary}{See the general documentation under \code{\link{buildmer-package}}}

\item{...}{Additional options to be passed to \code{gls}}
}
\description{
Use \code{buildmer} to fit generalized-least-squares models using \code{gls} from \code{nlme}
}
\examples{
library(buildmer)
library(nlme)
vowels$event <- with(vowels,interaction(participant,word))
m <- buildgls(f1 ~ timepoint*following,correlation=corAR1(form=~1|event),data=vowels)
}
\seealso{
\code{\link{buildmer-package}}
}
