% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildlme}
\alias{buildlme}
\title{Use buildmer to perform stepwise elimination of the fixed-effects part of mixed-effects models fit via lme() from nlme}
\usage{
buildlme(formula, data, random, cl = NULL, reduce.fixed = TRUE,
  direction = c("order", "backward"), crit = "LRT",
  calc.anova = TRUE, calc.summary = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{formula}{The model formula for the maximal model you would like to fit, if possible.}

\item{data}{The data to fit the models to.}

\item{random}{The random-effects specification for the model. This is not manipulated by buildlme() in any way!}

\item{cl}{An optional cluster object as returned by parallel::makeCluster() to use for parallelizing the evaluation of terms.}

\item{reduce.fixed}{Whether to reduce the fixed-effect structure.}

\item{direction}{The direction for stepwise elimination; possible options are `order' (order terms by their contribution to the model), `backward' (backward elimination), `forward' (forward elimination, implies `order'). The default is the combination `c('order','backward')', to first make sure that the model converges and to then perform backward elimination; other such combinations are perfectly allowed.}

\item{crit}{The criterion used to test terms for elimination. Possible options are `LRT', `AIC', and `BIC'.}

\item{calc.anova}{Whether to also calculate the ANOVA table for the final model after term elimination.}

\item{calc.summary}{Whether to also calculate the summary table for the final model after term elimination.}

\item{quiet}{Whether to suppress progress messages.}

\item{...}{Additional options to be passed to lme().}
}
\value{
A buildmer object containing the following slots:
\itemize{
\item model: the final model containing only the terms that survived elimination
\item p: the parameter list used in the various buildmer modules. Things of interest this list includes are, among others:
\itemize{
\item results: a dataframe containing the results of the elimination process
\item messages: any warning messages
} This information is also printed as part of the show() method.
\item summary: the model's summary, if `calc.summary=TRUE' was passed
\item anova: the model's anova, if `calc.anova=TRUE' was passed
}
}
\description{
Use buildmer to perform stepwise elimination of the fixed-effects part of mixed-effects models fit via lme() from nlme
}
\seealso{
buildmer
}
