% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildlme}
\alias{buildlme}
\title{Use \code{buildmer} to perform stepwise elimination of the fixed-effects part of mixed-effects models fit via \code{lme} from \code{nlme}}
\usage{
buildlme(formula, data = NULL, random, cl = NULL,
  direction = c("order", "backward"), crit = "LRT", include = NULL,
  calc.anova = TRUE, calc.summary = TRUE, ...)
}
\arguments{
\item{formula}{See the general documentation under \code{\link{buildmer-package}}}

\item{data}{See the general documentation under \code{\link{buildmer-package}}}

\item{random}{The random-effects specification for the model. This is not manipulated by \code{buildlme} in any way!}

\item{cl}{See the general documentation under \code{\link{buildmer-package}}}

\item{direction}{See the general documentation under \code{\link{buildmer-package}}}

\item{crit}{See the general documentation under \code{\link{buildmer-package}}}

\item{include}{See the general documentation under \code{\link{buildmer-package}}}

\item{calc.anova}{See the general documentation under \code{\link{buildmer-package}}}

\item{calc.summary}{See the general documentation under \code{\link{buildmer-package}}}

\item{...}{Additional options to be passed to \code{lme}}
}
\description{
Use \code{buildmer} to perform stepwise elimination of the fixed-effects part of mixed-effects models fit via \code{lme} from \code{nlme}
}
\examples{
library(buildmer)
m <- buildlme(Reaction ~ Days,data=lme4::sleepstudy,random=~Days|Subject)
}
\seealso{
\code{\link{buildmer-package}}
}
