% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\docType{class}
\name{buildmer-class}
\alias{buildmer-class}
\alias{mkBuildmer}
\title{Make a buildmer object}
\arguments{
\item{table}{A dataframe containing the results of the elimination process.}

\item{model}{The final model containing only the terms that survived elimination.}

\item{messages}{Any warning messages.}

\item{summary:}{The model's summary, if the model was built with `summary=TRUE'.}

\item{anova:}{The model's ANOVA, if the model was built with `anova=TRUE'.}
}
\description{
Make a buildmer object
}
\seealso{
buildmer
}
\keyword{buildmer}
