% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildmer}
\alias{buildmer}
\title{Construct and fit as complete a model as possible, and perform backward stepwise elimination using the change in deviance}
\usage{
buildmer(formula, data, family = gaussian, nAGQ = 1,
  adjust.p.chisq = TRUE, reduce.fixed = TRUE, reduce.random = TRUE,
  direction = c("backward", "forward", "null"), summary = FALSE,
  ddf = "Kenward-Roger", quiet = FALSE, verbose = 0, maxfun = 2e+05)
}
\arguments{
\item{formula}{The model formula; if you include random effects, use lme4 syntax for them.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or 'gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through.}

\item{nAGQ}{The number of Adaptive Gauss-Hermitian Quadrature points to use for glmer fits; default 1 (= Laplace approximation)}

\item{adjust.p.chisq}{Whether to adjust for overconservativity of the likelihood ratio test by dividing p-values by 2 (see Pinheiro & Bates 2000).}

\item{reduce.fixed}{Whether to reduce the fixed-effect structure.}

\item{reduce.random}{Whether to reduce the random-effect structure.}

\item{direction}{The direction for stepwise elimination; either 'forward' or 'backward' (default).}

\item{summary}{Whether to also calculate summaries for the final model after term elimination. This is rather pointless, except if you want to calculate degrees of freedom by Kenward-Roger approximation (default), in which case generating the summary (via lmerTest) will be very slow, and preparing the summary in advance can be advantageous.}

\item{ddf}{The adjustment to use in calculating the summary if summary=T and if lmerTest is available. Defaults to 'Kenward-Roger'.}

\item{quiet}{Whether to suppress progress messages.}

\item{verbose}{The verbosity level passed to (g)lmer fits.}

\item{maxfun}{The maximum number of iterations to allow for (g)lmer fits.}
}
\value{
A buildmer object containing the following slots:
\itemize{
\item table: a dataframe containing the results of the elimination process
\item model: the final model containing only the terms that survived elimination
\item messages: any warning messages
\item summary: the model's summary, if summary==TRUE
}
}
\description{
Construct and fit as complete a model as possible, and perform backward stepwise elimination using the change in deviance
}
\examples{
buildmer(Reaction~Days,list(Subject=~Days),lme4::sleepstudy)
}
\keyword{fit}

