% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{buildmer}
\alias{buildmer}
\title{Construct and fit as complete a model as possible, optionally reorder terms by their contribution to the deviance, and perform stepwise elimination using the change in deviance}
\usage{
buildmer(formula, data, family = gaussian, adjust.p.chisq = TRUE,
  reorder.terms = TRUE, cl = NULL, reduce.fixed = TRUE,
  reduce.random = TRUE, direction = "backward", calc.anova = TRUE,
  calc.summary = TRUE, ddf = "Wald", quiet = FALSE, ...)
}
\arguments{
\item{formula}{The model formula for the maximal model you would like to fit, if possible. Supports lme4 random effects and gamm4 smooth terms.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or `gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through.}

\item{adjust.p.chisq}{Whether to adjust for overconservativity of the likelihood ratio test by dividing p-values by 2 (see Pinheiro & Bates 2000).}

\item{reorder.terms}{Whether to reorder the terms by their contribution to the deviance before testing them.}

\item{cl}{An optional cluster object as returned by parallel::makeCluster() to use for parallelizing the evaluation of terms during the reordering step.}

\item{reduce.fixed}{Whether to reduce the fixed-effect structure.}

\item{reduce.random}{Whether to reduce the random-effect structure.}

\item{direction}{The direction for stepwise elimination; either `forward' or `backward' (default). Both or neither are also understood.}

\item{calc.anova}{Whether to also calculate the ANOVA table for the final model after term elimination. This is useful if you want to calculate degrees of freedom by Kenward-Roger approximation, in which case generating the ANOVA table (via lmerTest) will be very slow, and preparing the ANOVA in advance can be advantageous.}

\item{calc.summary}{Whether to also calculate the summary table for the final model after term elimination. This is useful if you want to calculate degrees of freedom by Kenward-Roger approximation (default), in which case generating the summary (via lmerTest) will be very slow, and preparing the summary in advance can be advantageous.}

\item{ddf}{The method used for calculating p-values if summary=TRUE. Options are `Wald' (default), `Satterthwaite' (if lmerTest is available), `Kenward-Roger' (if lmerTest and pbkrtest are available), and `lme4' (no p-values).}

\item{quiet}{Whether to suppress progress messages.}

\item{...}{Additional options to be passed to (g)lmer or gamm4. (They will also be passed to (g)lm in so far as they're applicable, so you can use arguments like `subset=...' and expect things to work. The single exception is the `control' argument, which is assumed to be meant only for (g)lmer and not for (g)lm, and will NOT be passed on to (g)lm.)}
}
\value{
A buildmer object containing the following slots:
\itemize{
\item model: the final model containing only the terms that survived elimination
\item p: the parameter list used in the various buildmer modules. Things of interest this list includes are, among others:
\itemize{
\item results: a dataframe containing the results of the elimination process
\item messages: any warning messages
}. This information is also printed as part of the show() method.
\item summary: the model's summary, if `calc.summary=TRUE' was passed
\item anova: the model's anova, if `calc.anova=TRUE' was passed
}
}
\description{
Construct and fit as complete a model as possible, optionally reorder terms by their contribution to the deviance, and perform stepwise elimination using the change in deviance
}
\examples{
buildmer(Reaction~Days+(Days|Subject),sleepstudy)
}
