% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular.R
\name{calcWald}
\alias{calcWald}
\title{Calculate p-values based on Wald z-scores}
\usage{
calcWald(table, i, sqrt = FALSE)
}
\arguments{
\item{table}{A coefficient table from a summary or anova output.}

\item{i}{The number of the column in that table containing the t-values.}

\item{sqrt}{Whether we're testing F values or t values (default).}
}
\value{
The table augmented with p-values.
}
\description{
Calculate p-values based on Wald z-scores
}
