% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular.R
\name{conv}
\alias{conv}
\title{Test a merMod or equivalent object for convergence}
\usage{
conv(model)
}
\arguments{
\item{model}{The model object to test.}
}
\value{
Whether the model converged or not.
}
\description{
Test a merMod or equivalent object for convergence
}
\keyword{convergence}
