% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{conv}
\alias{conv}
\title{Test an lme4 or equivalent object for convergence.}
\usage{
conv(model)
}
\arguments{
\item{model}{The model object to test.}
}
\description{
Test an lme4 or equivalent object for convergence.
}
\keyword{convergence}

