% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\docType{methods}
\name{diag,formula-method}
\alias{diag,formula-method}
\title{Diagonalize the random-effect covariance structure, possibly assisting convergence}
\usage{
\S4method{diag}{formula}(x)
}
\arguments{
\item{formula}{A model formula.}
}
\value{
The formula with all random-effect correlations forced to zero, per Pinheiro & Bates (2000).
}
\description{
Diagonalize the random-effect covariance structure, possibly assisting convergence
}
\keyword{covariance}
\keyword{diagonal}
\keyword{structure}
