% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{mer2tex}
\alias{mer2tex}
\title{Convert a buildmer (or compatible) model to LaTeX code}
\usage{
mer2tex(summary, aliases = list())
}
\arguments{
\item{summary}{The model (or its summary) to convert.}

\item{aliases}{A list of aliases translating summary terms to LaTeX code.}
}
\description{
Convert a buildmer (or compatible) model to LaTeX code
}
