% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{mer2tex}
\alias{A}
\alias{LaTeX}
\alias{aliases}
\alias{code.}
\alias{list}
\alias{mer2tex}
\alias{of}
\alias{summary}
\alias{terms}
\alias{to}
\alias{translating}
\title{Convert a summary to LaTeX code (biased towards vowel analysis)}
\usage{
mer2tex(summary, vowel = "", formula = F, diag = F, label = "",
  aliases = list(`(Intercept)` = "Intercept", Df1 = "$\\\\Updelta$F1", Df2 =
  "$\\\\Updelta$F2", ppn = "participants", word = "words", `2:` = "\\\\o:", `9y`
  = "\\\\oe y"))
}
\arguments{
\item{summary}{The summary to convert.}

\item{vowel}{The vowel you're analyzing.}

\item{formula}{The formula as used in your final lmer object.}

\item{diag}{Whether to include a note about a diagonal covariance structure having been assumed.}

\item{label}{The LaTeX label to put below your 'Results' caption.}
}
\description{
Convert a summary to LaTeX code (biased towards vowel analysis)
}
\keyword{LaTeX}

