% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexFuns.R
\name{mer2tex}
\alias{mer2tex}
\alias{A}
\alias{list}
\alias{of}
\alias{aliases}
\alias{translating}
\alias{summary}
\alias{terms}
\alias{to}
\alias{LaTeX}
\alias{code.}
\title{Convert a buildmer (or compatible) summary to LaTeX code (biased towards vowel analysis)}
\usage{
mer2tex(summary, vowel = "", formula = F, label = "", aliases = list())
}
\arguments{
\item{summary}{The summary to convert.}

\item{vowel}{The vowel you're analyzing.}

\item{formula}{The formula as used in your final lmer object.}

\item{label}{The LaTeX label to put below your 'Results' caption.}
}
\description{
Convert a buildmer (or compatible) summary to LaTeX code (biased towards vowel analysis)
}
\keyword{LaTeX}
