% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular.R
\name{stepwise}
\alias{stepwise}
\title{A simple interface to buildmer intended to mimic SPSS stepwise methods for term reordering and backward stepwise elimination}
\usage{
stepwise(formula, data, family = gaussian, ...)
}
\arguments{
\item{formula}{The model formula for the maximal model you would like to fit, if possible. Supports lme4 random effects and gamm4 smooth terms.}

\item{data}{The data to fit the models to.}

\item{family}{The error distribution to use. Only relevant for generalized models; if the family is empty or `gaussian', the models will be fit using lm(er), otherwise they will be fit using glm(er) with the specified error distribution passed through. Commonly-used options are either nothing/`gaussian' (linear regression), `binomial' (logistic regression), or `poisson' (loglin regression), although many other families exist (e.g. cloglog, ...).}

\item{...}{Additional parameters that override buildmer defaults, see 'buildmer'.}
}
\value{
A buildmer object, which you can use summary() on to get a summary of the final model, and elim() to get the list of eliminated terms.
}
\description{
A simple interface to buildmer intended to mimic SPSS stepwise methods for term reordering and backward stepwise elimination
}
\examples{
stepwise(Reaction~Days+(Days|Subject),sleepstudy)
}
\keyword{stepwise}
