% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\name{stepwise}
\alias{stepwise}
\title{Perform stepwise elimination like SPSS, without you having to make complex decisions.}
\usage{
stepwise(formula, data, family = gaussian, reorder.terms = T,
  direction = "backward")
}
\arguments{
\item{formula}{The model you would ideally like to fit (don't worry about convergence problems).}

\item{data}{The data you want to fit the model to.}

\item{family}{An optional argument denoting the type of error distribution to use. If you're doing logistic regression, specify 'family=binomial'. If you're doing Poisson regression, specify 'family=poisson'. If you just want a regular, linear regression, don't specify this argument at all.}

\item{direction}{If you want to use forward rather than backward elimination, specify 'direction=forward'. If you want both (forward followed by backward), specify 'direction=c("forward","backward")' (or, equivalently, 'direction=c("backward","forward")'). If you want backward elimination (the default), specify 'direction=backward', or don't specify this argument at all.}
}
\description{
Perform stepwise elimination like SPSS, without you having to make complex decisions.
}
\keyword{SPSS}
\keyword{stepwise,}

