% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmer.R
\docType{data}
\name{vowels}
\alias{vowels}
\title{Vowel data from a pilot study.}
\format{A standard data frame.}
\usage{
data(vowels)
}
\description{
Vowel data from a pilot study.
}
\examples{
#buildmer(f1 ~ vowel*timepoint*following + stress + information + (vowel*timepoint*following|participant),data=vowels,ddf='Satterthwaite',verbose=2,control=lmerControl(optCtrl=list(maxfun=250000))) #VERY slow
buildmer(f1 ~ vowel + timepoint + stress + following + information + vowel:timepoint + timepoint:following + vowel:following + vowel:timepoint:following + (1 + vowel + timepoint + following + vowel:timepoint + vowel:following + timepoint:following | participant),data=vowels,ddf='Satterthwaite',reduce.fixed=F,reduce.random=F,direction=NULL,start=c(0.567929,-0.0678053,-0.288065,-0.402232,-0.218813,-0.381153,-0.00170270,0.226449,-0.130479,0.304374,-0.0545927,-0.0540232,-0.0207310,0.106023,0.0406852,0.307277,0.227415,0.138530,0.227523,0.382057,0.235687,0.196619,-0.223459,-0.268986,-0.283158,-0.514559,-0.0852659,-0.0887947,-0.166943,-0.166356,0.0694036,0.104932,0.0315810,0.106377,0.572384,0.134674,0.0986679,-0.564443,-0.247899,-0.516127,0.0126328,0.0134181,0.156846,0.0735232,-0.295448,0.0760134,0.307895,-0.0320374,0.00561691,0.0263991,-0.0665717,-0.0271437,-0.375478,-0.225109,-0.138032,-0.167771,-0.314771,0.266541,0.143402,0.346499,-0.0542296,-0.0522401,-0.168416,-0.233742,-0.108847,0.169875,0.0587621,0.294170,-0.131189,0.00842303,0.00107080,-0.000140772,-0.000180338,-0.000391072,-0.000760899,0.000175467,0.000815001,0.000371120,0.00103038,7.76669e-05,-0.000361915,6.43889e-06,5.41626e-06,-6.29458e-06,5.28341e-06,-1.57044e-05,-1.21465e-05,-3.59313e-06,-1.08241e-05,3.79460e-06,-6.23642e-06,7.33735e-09,4.63050e-06,-1.05767e-06,8.90553e-06,5.19108e-06,1.14733e-06,4.10115e-06,-2.16305e-06,-3.79511e-07,6.61065e-08,2.25952e-07,-2.23288e-06,-1.86336e-07,1.99417e-07,-7.50342e-07,1.87314e-06,-5.57930e-07,2.81066e-08,1.74496e-06,-2.73743e-07,-1.26676e-07,4.62685e-07,-1.52272e-06,9.08249e-07,1.08874e-08,4.46392e-07,1.40782e-07,-7.20672e-09,1.00016e-07,-2.21006e-07,2.69678e-08,2.85674e-08,7.00216e-08,-2.39755e-07,7.03984e-08,8.71293e-08,4.31071e-08,-3.09939e-07,1.83222e-07,1.55808e-07,4.80558e-08,5.22165e-08,2.11040e-07,-1.53738e-07,3.32615e-07))
}
\keyword{datasets}

